Change file previews to use files, except for videos
While I would prefer not writing stuff and just using pipes, the use of
pipes has proved to be more error-prone. Sometimes, when moving between
many files quickly, the whole screen would get filled with garbage
output.
But, for some reason, trying to use a file for video thumbnails isn't as
reliable as piping the output from ffmpegthumbnailer, as sometimes the
thumbnail would not get displayed if using a file.